+++ CHANGELOG.rst
@@ -58,6 +58,9 @@ Runtime Behavior Changes
 Bug Fixes
 ~~~~~~~~~
 
+* finagle-http: Finagle now properly sets the `Transport.peerCertificate` local context
+  when using HTTP/2. ``PHAB_ID=D324392``
+
 * finagle-zipkin-core: Fix a race condition which could cause a span to get logged
   missing some annotations. ``PHAB_ID=D319367``